requirements.txt
requests
pandas
beautifulsoup4
python-dotenv
openai
playwright
gradio
Flask